feat: #62 DailyTodoHistoryService 히스토리 저장 로직 변경
